sqlclustersettings: Refactor sql cluster settings to a separate package
This commit creates sqlclustersetting package and moves some sql
cluster settings there. The hope is that this new package will from now
on host sql related cluster settings.

Release note: None
