+++ pkg/ccl/backupccl/BUILD.bazel
@@ -120,6 +120,7 @@ go_library(
         "//pkg/sql/sem/eval",
         "//pkg/sql/sem/tree",
         "//pkg/sql/sessiondata",
+        "//pkg/sql/sqlclustersettings",
         "//pkg/sql/sqlerrors",
         "//pkg/sql/stats",
         "//pkg/sql/syntheticprivilege",
@@ -285,6 +286,7 @@ go_test(
         "//pkg/sql/sem/eval",
         "//pkg/sql/sem/tree",
         "//pkg/sql/sessiondata",
+        "//pkg/sql/sqlclustersettings",
         "//pkg/sql/sqlliveness/slinstance",
         "//pkg/sql/stats",
         "//pkg/storage",
