+++ pkg/sql/sqlclustersettings/BUILD.bazel
@@ -1,0 +1,9 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "sqlclustersettings",
+    srcs = ["clustersettings.go"],
+    importpath = "github.com/cockroachdb/cockroach/pkg/sql/sqlclustersettings",
+    visibility = ["//visibility:public"],
+    deps = ["//pkg/settings"],
+)
